reUsable components with component composition
