+++ src/components/movies/WatchedList.jsx
@@ -1,0 +1,30 @@
+import React from 'react'
+
+const WatchedList = ({watched}) => {
+  return (
+    <ul className="list">
+            {watched.map((movie) => (
+              <li key={movie.imdbID}>
+                <img src={movie.Poster} alt={`${movie.Title} poster`} />
+                <h3>{movie.Title}</h3>
+                <div>
+                  <p>
+                    <span>⭐️</span>
+                    <span>{movie.imdbRating}</span>
+                  </p>
+                  <p>
+                    <span>🌟</span>
+                    <span>{movie.userRating}</span>
+                  </p>
+                  <p>
+                    <span>⏳</span>
+                    <span>{movie.runtime} min</span>
+                  </p>
+                </div>
+              </li>
+            ))}
+          </ul>
+  )
+}
+
+export default WatchedList
