+++ src/components/movies/WatchedSummary.jsx
@@ -1,0 +1,29 @@
+import React from 'react'
+
+const WatchedSummary = ({watched,avgImdbRating,avgUserRating, avgRuntime}) => {
+  return (
+    <div className="summary">
+            <h2>Movies you watched</h2>
+            <div>
+              <p>
+                <span>#️⃣</span>
+                <span>{watched.length} movies</span>
+              </p>
+              <p>
+                <span>⭐️</span>
+                <span>{avgImdbRating}</span>
+              </p>
+              <p>
+                <span>🌟</span>
+                <span>{avgUserRating}</span>
+              </p>
+              <p>
+                <span>⏳</span>
+                <span>{avgRuntime} min</span>
+              </p>
+            </div>
+          </div>
+  )
+}
+
+export default WatchedSummary
